Move search paths resolving methods to a separate module
