+++ packages/search-in-workspace/src/node/ripgrep-search-in-workspace-server.ts
@@ -20,8 +20,9 @@ import { FileUri } from '@theia/core/lib/node/file-uri';
 import URI from '@theia/core/lib/common/uri';
 import { inject, injectable } from '@theia/core/shared/inversify';
 import { SearchInWorkspaceServer, SearchInWorkspaceOptions, SearchInWorkspaceResult, SearchInWorkspaceClient, LinePreview } from '../common/search-in-workspace-interface';
-import * as fs from '@theia/core/shared/fs-extra';
 import * as path from 'path';
+import { isRelativeToBaseDirectory, pushIfNotIncluded } from './ripgrep-search-in-workspace-helper';
+import { resolveSearchPathsFromIncludes } from './ripgrep-search-paths-resolver';
 
 export const RgPath = Symbol('RgPath');
 
@@ -40,7 +41,8 @@ function bytesOrTextToString(obj: IRgBytesOrText): string {
 
 /**
  * Joins the given root and pattern to form an absolute path
- * as long as the pattern is in relative form e.g. './foo'.
+ * as long as the pattern is in relative form.
+ * E.g. './foo' becomes '${root}/foo'
  */
 function relativeToAbsolutePattern(root: string, pattern: string): string {
     if (!isRelativeToBaseDirectory(pattern)) {
@@ -48,65 +50,6 @@ function relativeToAbsolutePattern(root: string, pattern: string): string {
         return pattern;
     }
     return path.join(root, pattern);
-}
-
-function isRelativeToBaseDirectory(filePath: string): boolean {
-    const normalizedPath = filePath.replace(/\\/g, '/');
-    return normalizedPath.startsWith('./');
-}
-
-/**
- * Attempts to build a valid absolute file or directory from the given pattern and root folder.
- * e.g. /a/b/c/foo/** to /a/b/c/foo, or './foo/**' to '${root}/foo'.
- *
- * @returns the valid path if found existing in the file system.
- */
-function resolveIncludeFolderFromGlob(root: string, pattern: string): string | undefined {
-    const patternBase = stripGlobSuffix(pattern);
-
-    if (!path.isAbsolute(patternBase) && !isRelativeToBaseDirectory(patternBase)) {
-        // The include pattern is not referring to a single file / folder, i.e. not to be converted
-        // to include folder.
-        return undefined;
-    }
-
-    const targetPath = path.isAbsolute(patternBase) ? patternBase : path.join(root, patternBase);
-
-    if (fs.existsSync(targetPath)) {
-        return targetPath;
-    }
-
-    return undefined;
-}
-
-/**
- * Removes a glob suffix from a given pattern (e.g. /a/b/c/**)
- * to a directory path (/a/b/c).
- *
- * @returns the path without the glob suffix,
- * else returns the original pattern.
- */
-function stripGlobSuffix(pattern: string): string {
-    const pathParsed = path.parse(pattern);
-    const suffix = pathParsed.base;
-
-    return suffix === '**' ? pathParsed.dir : pattern;
-}
-
-/**
- * Push an item to an existing string array only if the item is not already included.
- * If the given array is undefined it creates a new one with the given item as the first entry.
- */
-function pushIfNotIncluded(containerArray: string[] | undefined, item: string): string[] {
-    if (!containerArray) {
-        return [item];
-    }
-
-    if (!containerArray.includes(item)) {
-        containerArray.push(item);
-    }
-
-    return containerArray;
 }
 
 type IRgMessage = IRgMatch | IRgBegin | IRgEnd;
@@ -260,7 +203,7 @@ export class RipgrepSearchInWorkspaceServer implements SearchInWorkspaceServer {
         // we'll use to parse the lines.
         const searchId = this.nextSearchId++;
         const rootPaths = rootUris.map(root => FileUri.fsPath(root));
-        const searchPaths: string[] = this.resolveSearchPathsFromIncludes(rootPaths, opts);
+        const searchPaths: string[] = resolveSearchPathsFromIncludes(rootPaths, opts);
         this.includesExcludesToAbsolute(searchPaths, opts);
         const rgArgs = this.getArgs(opts);
         // if we use matchWholeWord we use regExp internally,
@@ -439,55 +382,6 @@ export class RipgrepSearchInWorkspaceServer implements SearchInWorkspaceServer {
                 opts!.exclude = updatedPatterns;
             }
         });
-    }
-
-    /**
-     * The default search paths are set to be the root paths associated to a workspace
-     * however the search scope can be further refined with the include paths available in the search options.
-     * This method will replace the searching paths to the ones specified in the 'include' options but as long
-     * as the 'include' paths can be successfully validated as existing.
-     *
-     * Therefore the returned array of paths can be either the workspace root paths or a set of validated paths
-     * derived from the include options which can be used to perform the search.
-     *
-     * Any pattern that resulted in a valid search path will be removed from the 'include' list as it is
-     * provided as an equivalent search path instead.
-     */
-    protected resolveSearchPathsFromIncludes(rootPaths: string[], opts: SearchInWorkspaceOptions | undefined): string[] {
-        if (!opts || !opts.include) {
-            return rootPaths;
-        }
-
-        const includesAsPaths = this.resolvePatternToPathsMap(opts.include, rootPaths);
-        const patternPaths = Array.from(includesAsPaths.keys());
-
-        // Remove file patterns that were successfully translated to search paths.
-        opts.include = opts.include.filter(item => !patternPaths.includes(item));
-
-        return includesAsPaths.size > 0 ? [].concat.apply([], Array.from(includesAsPaths.values())) : rootPaths;
-    }
-
-    /**
-     * Attempts to resolve valid file paths from a given list of patterns.
-     * The given search paths are used to try resolving relative path patterns to an absolute path.
-     * The resulting map will include all patterns associated to its equivalent file paths.
-     * The given patterns that are not successfully mapped to paths are not included.
-     */
-    protected resolvePatternToPathsMap(patterns: string[], searchPaths: string[]): Map<string, string[]> {
-        const patternToPathMap = new Map<string, string[]>();
-
-        patterns.forEach(pattern => {
-            searchPaths.forEach(root => {
-                const foundPath = resolveIncludeFolderFromGlob(root, pattern);
-
-                if (foundPath) {
-                    const pathArray = patternToPathMap.get(pattern);
-                    patternToPathMap.set(pattern, pushIfNotIncluded(pathArray, foundPath));
-                }
-            });
-        });
-
-        return patternToPathMap;
     }
 
     /**
